added missing translations and moved interface declarations
