+++ packages/esm-outpatient-app/src/current-visit/current-visit.resource.ts
@@ -1,5 +1,5 @@
 import useSWR from 'swr';
-import { openmrsFetch, OpenmrsResource, Visit } from '@openmrs/esm-framework';
+import { openmrsFetch, Visit } from '@openmrs/esm-framework';
 
 export function useVisit(visitUuid: string) {
   const customRepresentation =
@@ -26,91 +26,6 @@ export function useVisit(visitUuid: string) {
     isLoading: !data && !error,
     isValidating,
   };
-}
-
-export interface Note {
-  concept: OpenmrsResource;
-  note: string;
-  provider: {
-    name: string;
-    role: string;
-  };
-  time: string;
-}
-
-export interface DiagnosisItem {
-  diagnosis: string;
-}
-
-export interface Encounter {
-  uuid: string;
-  encounterDatetime: string;
-  encounterProviders: Array<{
-    uuid: string;
-    display: string;
-    encounterRole: {
-      uuid: string;
-      display: string;
-    };
-    provider: {
-      uuid: string;
-      person: {
-        uuid: string;
-        display: string;
-      };
-    };
-  }>;
-  encounterType: {
-    uuid: string;
-    display: string;
-  };
-  obs: Array<Observation>;
-  form: OpenmrsResource;
-  patient: OpenmrsResource;
-}
-
-export interface Observation {
-  uuid: string;
-  concept: {
-    uuid: string;
-    display: string;
-    conceptClass: {
-      uuid: string;
-      display: string;
-    };
-  };
-  display: string;
-  groupMembers: null | Array<{
-    uuid: string;
-    concept: {
-      uuid: string;
-      display: string;
-    };
-    value: {
-      uuid: string;
-      display: string;
-    };
-  }>;
-  value: any;
-  obsDatetime: string;
-}
-
-export interface PatientVitals {
-  systolic?: number;
-  diastolic?: number;
-  pulse?: number;
-  temperature?: number;
-  oxygenSaturation?: number;
-  height?: number;
-  weight?: number;
-  bmi?: number | null;
-  respiratoryRate?: number;
-  muac?: number;
-  provider?: {
-    name: string;
-    role: string;
-  };
-  time?: string;
 }
 
 export function calculateBMI(weight: number, height: number): number {
