+++ packages/esm-outpatient-app/src/current-visit/visit-details/triage-note.scss
@@ -34,6 +34,5 @@
 }
 
 .notification {
-  background-color: var(--open-mrs-danger-red-60);
   fill: $danger;
 }
