+++ packages/esm-outpatient-app/translations/en.json
@@ -92,5 +92,24 @@
   "waitingFor": "Waiting for:",
   "waitTime": "Wait time (mins)",
   "youngest": "Youngest first",
-  "backToScheduledVisits": "Back To Scheduled Visits"
+  "backToScheduledVisits": "Back To Scheduled Visits",
+  "scheduledToday": "Scheduled for today",
+  "onTime": "On time",
+  "tirageNotYetCompleted": "Triage has not yet been completed",
+  "vitalsNotRecordedForVisit": "Vitals has not been recorded for this patient for this visit",
+  "kg": "kg",
+  "weight": "Weight",
+  "kg/m²": "kg / m²",
+  "bmi": "Bmi",
+  "cm": "cm",
+  "height": "Height",
+  "/min": "/ min",
+  "rRate": "R. Rate",
+  "sp02": "Sp02",
+  "bpm": "bpm",
+  "heartRate": "Heart rate",
+  "mmHg": "mmHg",
+  "bp": "Bp",
+  "°C": "°C",
+  "temperature": "Temperature"
 }
